refactor: use view model function to make request
